old_version/web/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pyjwt==2.6.0
bcrypt==4.0.1
redis==4.5.1
python-jose>=3.3.0
python-dotenv==1.0.1
email-validator>=2.1.0.post1